+++ my-app/src/Game/Components/interfaces/ItemProps.ts
@@ -1,4 +1,4 @@
-import { Item } from "../../Core/StartData/StartData";
+import {Item} from "../../Core/Items/Item";
 
 export interface ItemProps {
     key: number,
