+++ my-app/src/Game/Components/interfaces/ProjectileProps.ts
@@ -1,4 +1,5 @@
-import { HeroProjectile } from '../../Core/StartData/StartData';
+import {HeroProjectile} from "../../Core/HeroProjectile/HeroProjectile";
+
 export interface ProjectileProps {
     key : number, 
     projectileData : HeroProjectile
